integrating w image gen changes
